scheduler/requirements.txt
pyyaml==6.0.0
python-dotenv==0.20.0
backoff==2.1.2
orjson==3.8.6
psycopg2-binary==2.9.3
pydantic==1.9.1
requests==2.28.0
